couvrance des tests à 100%  et correction du bug de __main.py

+++ olympics/__main__.py
@@ -5,33 +5,34 @@ import argparse
 from . import cli
 
 parser = argparse.ArgumentParser(
-    prog='olympics',
-    description='Display various information about Olympics results',
+    prog="olympics",
+    description="Display various information about Olympics results",
 )
 parser.add_argument(
-    'command',
-    help='command to launch',
-    choices=('countries', 'collective', 'individual'),
+    "command",
+    help="command to launch",
+    choices=("countries", "collective", "individual"),
 )
 parser.add_argument(
-    '--top',
-    help='number of top elements to display',
-    type=float,
+    "--top",
+    help="number of top elements to display",
+    type=int,  # correction du bug
     default=10,
 )
+
 
 def main(argv=None):
     args = parser.parse_args(argv)
     if (top := int(args.top)) <= 0:
-        raise argparse.ArgumentTypeError(f'{top} is not a positive number')
+        raise argparse.ArgumentTypeError(f"{top} is not a positive number")
     match args.command:
-        case 'countries':
+        case "countries":
             cli.top_countries(top)
-        case 'collective':
+        case "collective":
             cli.top_collective(top)
-        case 'individual':
+        case "individual":
             cli.top_individual(top)
 
 
-if __name__ == '__main__':  # pragma: no cover
+if __name__ == "__main__":  # pragma: no cover
     main()
